requirements.txt
allure-behave==2.8.16
allure-pytest==2.8.16
allure-python-commons==2.8.16
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
behave==1.2.6
cassandra-driver==3.24.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
geomet==0.2.1.post1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
MouseInfo==0.1.3
packaging==20.4
parse==1.15.0
parse-type==0.5.2
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pytest==5.4.3
PyTweening==1.0.3
requests==2.24.0
selenium==3.141.0
six==1.15.0
snowballstemmer==2.0.0
toml==0.10.1
urllib3==1.25.9
wcwidth==0.2.5
wrapt==1.12.1
cx_Oracle==8.0.1